+++ src/Exceptions/InvalidRendererException.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Magdicom;
+namespace Magdicom\Exceptions;
 
+use Magdicom\Renderer;
 use RuntimeException;
 
 class InvalidRendererException extends RuntimeException