+++ src/Exceptions/MissingRendererException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Magdicom;
+namespace Magdicom\Exceptions;
 
 use RuntimeException;