+++ src/Processors/LastProcessor.php
@@ -2,16 +2,20 @@
 
 declare(strict_types=1);
 
-namespace Magdicom\Processor;
+namespace Magdicom\Processors;
 
 use Magdicom\ProcessingContext;
 use Magdicom\ResultProcessor;
 
 /** @implements ResultProcessor<mixed, mixed> */
-class FirstProcessor implements ResultProcessor
+class LastProcessor implements ResultProcessor
 {
     public function process(array $results, ProcessingContext $context): mixed
     {
-        return $results[0] ?? null;
+        if ($results === []) {
+            return null;
+        }
+
+        return $results[array_key_last($results)];
     }
 }